Extract parabol tasks panel to separate component
